requirements.txt
black==22.8.0
click==8.1.3
flake8==5.0.4
mccabe==0.7.0
mypy-extensions==0.4.3
numpy==1.23.3
pathspec==0.10.1
platformdirs==2.5.2
pycodestyle==2.9.1
pyflakes==2.5.0
PyQt6==6.3.1
PyQt6-Qt6==6.3.1
PyQt6-sip==13.4.0
pyqtgraph==0.12.4
pyserial==3.5
screeninfo==0.8.1
tomli==2.0.1
